\name{RLI_SGPt_Baseline_Matrices}
\alias{RLI_SGPt_Baseline_Matrices}
\docType{data}
\title{Baseline matrices for RLI baseline SGP analyses}
\description{
An environment containing collections of baseline matrices used historically during RLI baseline analyses. Matrices are updated (added to) at the end of the Fall, Winter, Spring windows.
Matrices are used as part of the \code{rliSGP} function call. The environment currently contains matrices associated with the following window based analyses:

\describe{
    \item{\code{RLI_SGPt_Baseline_Matrices_2017_2018.1}}{Winter, 2019-2020 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2017_2018.1}}{Spring, 2023-2024 SGPt analyses}
}
}

\source{Analyses of RLI Star data}
\keyword{datasets}
