\name{RLIdata-package}
\alias{RLIdata-package}
\docType{package}
\title{Historical baseline coefficient matrices for RLI SGPt analyses}
\description{
	The package RLIdata contains historical baseline coefficient matrices for RLI SGPt analyses.
}

\details{
\tabular{ll}{
Package: \tab RLIdata\cr
Type: \tab Package\cr
Version: \tab 1.0-0.0\cr
Date: \tab 2015-10-3\cr
License: \tab GPL-3\cr
LazyLoad: \tab Yes\cr
LazyData: \tab Yes\cr
}
}

\keyword{package}
