\name{RLI_SGPt_Baseline_Matrices}
\alias{RLI_SGPt_Baseline_Matrices}
\docType{data}
\title{Baseline matrices for RLI baseline SGP analyses}
\description{
An environment containing collections of baseline matrices used historically during RLI baseline analyses. Matrices are updated (added to) at the end of the Fall, Winter, Spring windows.
Matrices are used as part of the \code{rliSGP} function call. The environment currently contains matrices associated with the following window based analyses:

\describe{
    \item{\code{RLI_SGPt_Baseline_Matrices_2013_2014.1}}{Fall, 2013-2014 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2013_2014.2}}{Winter, 2013-2014 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2013_2014.3}}{Spring, 2013-2014 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2014_2015.1}}{Fall, 2014-2015 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2014_2015.2}}{Winter, 2014-2015 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2014_2015.3}}{Spring, 2014-2015 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2015_2016.1}}{Fall, 2015-2016 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2015_2016.2}}{Winter, 2015-2016 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2015_2016.3}}{Spring, 2015-2016 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2016_2017.1}}{Fall, 2016-2017 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2016_2017.2}}{Winter, 2016-2017 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2016_2017.3}}{Spring, 2016-2017 SGPt analyses}
    \item{\code{RLI_SGPt_Baseline_Matrices_2017_2018.1}}{Fall, 2017-2018 SGPt analyses}
}
}

\source{Analyses of RLI Star data}
\keyword{datasets}
