\name{RLImatrices-package}
\alias{RLImatrices-package}
\docType{package}
\title{Historical baseline coefficient matrices for RLI SGPt analyses}
\description{
	The package RLImatrices contains historical baseline coefficient matrices for RLI SGPt analyses.
}

\details{
\tabular{ll}{
Package: \tab RLImatrices\cr
Type: \tab Package\cr
Version: \tab 18.9-1.0\cr
Date: \tab 2025-3-15\cr
License: \tab MIT\cr
LazyLoad: \tab Yes\cr
LazyData: \tab Yes\cr
}
}

\keyword{package}
