\name{RLImatrices-package}
\alias{RLImatrices-package}
\docType{package}
\title{Historical baseline coefficient matrices for RLI SGPt analyses}
\description{
	The package RLImatrices contains historical baseline coefficient matrices for RLI SGPt analyses.
}

\details{
\tabular{ll}{
Package: \tab RLImatrices\cr
Type: \tab Package\cr
Version: \tab 14.0-0.0\cr
Date: \tab 2020-5-28\cr
License: \tab GPL-3\cr
LazyLoad: \tab Yes\cr
LazyData: \tab Yes\cr
}
}

\keyword{package}
